refactor(cli): route commands through focused modules
